# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: BharatStudio Alerts API
  version: 1.0.0-draft
  description: >-
    Draft v1 REST contract for BharatStudio Alerts and Companion. This contract
    is not approved for production until L01 review and verification pass.
servers:
  - url: https://api.example.invalid
    description: Placeholder only; production domain is not yet approved.
...
paths:
...
  /v1/channels:
    post:
      tags: [Creator]
      operationId: createChannel
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/channels/{channelId}:
    get:
      tags: [Creator]
      operationId: getChannelDetails
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel details visible to the authenticated member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Creator]
      operationId: updateChannelDetails
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
      responses:
        '200':
          description: Updated channel details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/channels/{channelId}/config:
    get:
      tags: [Configuration]
      operationId: getChannelConfig
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Versioned channel alert configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Configuration]
      operationId: updateChannelConfig
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelConfigRequest'
      responses:
        '200':
          description: New immutable configuration version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/channels/{channelId}/queues:
    get:
      tags: [Queues]
      operationId: listQueues
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Queues visible to the authenticated member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Queues]
      operationId: createQueue
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQueueRequest'
      responses:
        '201':
          description: Queue created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/channels/{channelId}/queues/{queueId}:
    patch:
      tags: [Queues]
      operationId: updateQueue
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/QueueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQueueRequest'
      responses:
        '200':
          description: Updated queue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/channels/{channelId}/bindings:
    get:
      tags: [Queues]
      operationId: listQueueBindings
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Source-to-queue routing bindings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Queues]
      operationId: createQueueBinding
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBindingRequest'
      responses:
        '201':
          description: Binding created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueBinding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/channels/{channelId}/bindings/{bindingId}:
    patch:
      tags: [Queues]
      operationId: updateQueueBinding
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - name: bindingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBindingRequest'
      responses:
        '200':
          description: Binding updated. Accepted delivery snapshots are unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueBinding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    QueueId:
      name: queueId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema: { type: integer, minimum: 1 }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    viewerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-viewer-session-token
    overlayToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-overlay-token
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: OIDC-service-identity-token
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated identity lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: Request conflicts with current resource/version state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [schemaVersion, errorCode, message, traceId]
      properties:
        schemaVersion: { type: string, const: v1 }
        errorCode: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        traceId: { type: string, minLength: 1, maxLength: 128 }
        retryable: { type: boolean }
...
    CreateChannelRequest:
      type: object
      additionalProperties: false
      required: [handle, displayName]
      properties:
        handle: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
        displayName: { type: string, minLength: 1, maxLength: 120 }
    UpdateChannelRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName: { type: string, minLength: 1, maxLength: 120 }
        acceptingTips: { type: boolean }
        featuredConsent: { type: boolean }
        handle: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
    ChannelDetails:
      type: object
      additionalProperties: false
      required: [schemaVersion, channelId, handle, displayName, acceptingTips, publicConfigVersion, featuredConsent]
      properties:
        schemaVersion: { type: string, const: v1 }
        channelId: { type: string, format: uuid }
        handle: { type: string }
        displayName: { type: string, maxLength: 120 }
        acceptingTips: { type: boolean }
        publicConfigVersion: { type: integer, minimum: 1 }
        featuredConsent: { type: boolean }
        role: { $ref: '#/components/schemas/ChannelMembership/properties/role' }
    ChannelConfig:
      type: object
      additionalProperties: false
      required: [schemaVersion, channelId, version, values, effectiveAt]
      properties:
        schemaVersion: { type: string, const: v1 }
        channelId: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        values: { type: object, additionalProperties: true }
        effectiveAt: { type: string, format: date-time }
    UpdateChannelConfigRequest:
      type: object
      additionalProperties: false
      required: [values]
      properties:
        values: { type: object, additionalProperties: true }
    Queue:
      type: object
      additionalProperties: false
      required: [schemaVersion, queueId, channelId, name, paused, active]
      properties:
        schemaVersion: { type: string, const: v1 }
        queueId: { type: string, format: uuid }
        channelId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 80 }
        paused: { type: boolean }
        active: { type: boolean }
    QueueList:
      type: object
      additionalProperties: false
      required: [schemaVersion, queues]
      properties:
        schemaVersion: { type: string, const: v1 }
        queues:
          type: array
          items: { $ref: '#/components/schemas/Queue' }
    CreateQueueRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
    UpdateQueueRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        paused: { type: boolean }
        active: { type: boolean }
...
    QueueBinding:
      type: object
      additionalProperties: false
      required: [schemaVersion, bindingId, channelId, queueId, sourceType, sourceId, allowDuplicates, priority]
      properties:
        schemaVersion: { type: string, const: v1 }
        bindingId: { type: string, format: uuid }
        channelId: { type: string, format: uuid }
        queueId: { type: string, format: uuid }
        sourceType: { type: string, enum: [payment, manual, companion] }
        sourceId: { type: string, minLength: 1, maxLength: 160 }
        allowDuplicates: { type: boolean }
        priority: { type: integer, minimum: 0, maximum: 100000 }
        overrideValues: { $ref: '#/components/schemas/BindingOverrideValues' }
        active: { type: boolean }
    BindingList:
      type: object
      additionalProperties: false
      required: [schemaVersion, bindings]
      properties:
        schemaVersion: { type: string, const: v1 }
        bindings:
          type: array
          items: { $ref: '#/components/schemas/QueueBinding' }
    CreateBindingRequest:
      type: object
      additionalProperties: false
      required: [queueId, sourceType, sourceId, allowDuplicates, priority]
      properties:
        queueId: { type: string, format: uuid }
        sourceType: { type: string, enum: [payment, manual, companion] }
        sourceId: { type: string, minLength: 1, maxLength: 160 }
        allowDuplicates: { type: boolean }
        priority: { type: integer, minimum: 0, maximum: 100000 }
        overrideValues: { $ref: '#/components/schemas/BindingOverrideValues' }
    UpdateBindingRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        allowDuplicates: { type: boolean }
        priority: { type: integer, minimum: 0, maximum: 100000 }
        overrideValues: { $ref: '#/components/schemas/BindingOverrideValues' }
        active: { type: boolean }
...